Code review 1 - move objects to different places
